+++ src/game/game_syscalls.c
@@ -27,8 +27,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 #include "game_local.h"
 
-static int (QDECL *syscall)(intptr_t arg, ...) = 
-	(int (QDECL *)(intptr_t, ...)) -1;
+static intptr_t (QDECL *syscall)(intptr_t arg, ...) = 
+	(intptr_t (QDECL *)(intptr_t, ...)) -1;
 
 void DLLEXPORT dllEntry(int(QDECL *syscallptr)(intptr_t arg, ...))
 {
